crypt/requirements.txt
bitarray==3.3.1
crypto==1.4.1
flask-cors==5.0.1
gmpy2==2.2.1
pip==25.0.1
pycryptodome==3.22.0
pycryptodomex==3.22.0
python-dotenv==1.1.0
